timers/dispatching: always insert timers into the DB
the optimization makes sense for reminders, but in this case,
suppose the user sets up a 5 second message expiration time.
If the bot shuts down, potentially many messages that are still pending deletion
would be lost with this optimization in place.
So we must always insert into the db so that pending timers with short expiries still get dispatched
when the bot starts up.